@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>WinExe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net7.0-windows10.0.17763.0</TargetFramework>
     <UseWPF>true</UseWPF>
     <UseWindowsForms>true</UseWindowsForms>
     <Configurations>Debug;Release;ReleaseToPublic</Configurations>
@@ -25,13 +25,13 @@
   <ItemGroup>
     <PackageReference Include="AvalonEdit" Version="6.1.3.50" />
     <PackageReference Include="CefSharp.Wpf.NETCore" Version="98.1.210" />
-    <PackageReference Include="DotNetZip" Version="1.16.0" />
+	<!--PackageReference Include="DotNetZip" Version="1.16.0" /-->
     <PackageReference Include="GameAnalytics.Net.Core.SDK" Version="3.3.5" />
     <PackageReference Include="HandyControl" Version="3.2.0" />
     <PackageReference Include="LibGit2Sharp" Version="0.27.0-preview-0175" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
     <PackageReference Include="Sentry" Version="3.14.1" />
-    <PackageReference Include="SpicyTaco.AutoGrid" Version="1.2.29" />
+	<!--PackageReference Include="SpicyTaco.AutoGrid" Version="1.2.29" /-->
   </ItemGroup>
 
   <ItemGroup>
